many implementations, we are halfway there
