+++ Bootstrap/AppDelegate.m
@@ -24,6 +24,8 @@ UITextView* logView=nil;
         [logView setText:[logView.text stringByAppendingString:[NSString stringWithFormat:@"%@\n",text]]];
         if(logView.contentSize.height >= logView.bounds.size.height)
             [logView setContentOffset:CGPointMake(0, logView.contentSize.height - logView.bounds.size.height) animated:YES];
+        const char *cString = [text UTF8String];
+        printf("%s", cString);
     });
 }
