+++ Bootstrap/AppView.swift
@@ -1,0 +1,14 @@
+//
+//  AppView.swift
+//  Bootstrap
+//
+//  Created by haxi0 on 31.12.2023.
+//
+
+import SwiftUI
+
+struct AppView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
