+++ Bootstrap/Bootstrap-Bridging-Header.h
@@ -2,3 +2,5 @@
 //  Use this file to import your target's public headers that you would like to expose to Swift.
 //
 
+#include "bootstrap.h"
+#include "bootstrapFr.h"
