+++ Bootstrap/Classes/Log.swift
