+++ Bootstrap/Private Headers/LSApplicationProxy.h
@@ -1,0 +1,12 @@
+//
+//  LSApplicationProxy.h
+//  Bootstrap
+//
+//  Created by haxi0 on 31.12.2023.
+//
+
+#ifndef LSApplicationProxy_h
+#define LSApplicationProxy_h
+
+
+#endif /* LSApplicationProxy_h */
