+++ Bootstrap/bootstrapFr.h
@@ -1,0 +1,16 @@
+//
+//  bootstrapFr.h
+//  Bootstrap
+//
+//  Created by haxi0 on 31.12.2023.
+//
+
+#ifndef bootstrapFr_h
+#define bootstrapFr_h
+
+void bootstrapFr(void);
+void unbootstrapFr(void);
+void respringFr(void);
+void rebuildappsFr(void);
+
+#endif /* bootstrapFr_h */
